Update and rename main.py to coin_toss.py
updated the source code to do the experiment for a user-defined number of coins
